Better adjacency logic.
This works correctly for the 'flattened' mesh, but it appears that the
sphere mesh doesn't actually encode the fact that the poles triangles
all meet at the same vertex. Bleh. Don't know if there's a way to have
proper geometry meshes, or if THREE.js even has a concept a closed
surface.